requirements.txt
numpy
requests
spotipy
pandas
pillow
python-dotenv
selenium
pytest
flask